Rename ReportActionsView to ChatReportActionsView and strip dead branches
